Cleaned up the code somewhat and better structured the tasks.  All tasks now use in-memory large datasets whenever necessary.

+++ src/net/sviglas/meridian/task/IndexRange.java
@@ -4,16 +4,16 @@ import java.util.Iterator;
 
 import net.sviglas.util.Pair;
 
-public class IndexRange extends Range<Integer> {    
-    private int low;
-    private int high;
-    private int threshold;
+public class IndexRange extends Range<Long> {
+    private long low;
+    private long high;
+    private long threshold;
 
-    public IndexRange(int l, int h) {
+    public IndexRange(long l, long h) {
         this(l, h, 1);
     }
 
-    public IndexRange(int l, int h, int t) {
+    public IndexRange(long l, long h, long t) {
         low = l;
         high = h;
         threshold = t;
@@ -23,31 +23,31 @@ public class IndexRange extends Range<Integer> {
         return high - low <= threshold;
     }
 
-    public Pair<Range<Integer>, Range<Integer>> split() {
-        int mid = (low + high) / 2;
+    public Pair<Range<Long>, Range<Long>> split() {
+        long mid = (low + high) / 2;
         return new Pair<>(new IndexRange(low, mid, threshold),
                 new IndexRange(mid, high, threshold));
     }
 
-    public Iterator<Integer> iterator() {
+    public Iterator<Long> iterator() {
         return new IndexIterator(low, high);
     }
 
-    public Integer begin() {
+    public Long begin() {
         return low;
     }
 
-    public Integer end() {
+    public Long end() {
         return high;
     }
 
-    public class IndexIterator implements Iterator<Integer> {
-        private int low;
-        private int high;
-        private int current;
-        public IndexIterator(int l, int h) { low = l; high = h; current = l; }
+    public class IndexIterator implements Iterator<Long> {
+        private long low;
+        private long high;
+        private long current;
+        public IndexIterator(long l, long h) { low = l; high = h; current = l; }
         public boolean hasNext() { return current++ < high; }
-        public Integer next() { return current; }
+        public Long next() { return current; }
         public void remove() {}
     }
 }
